requirements.txt
colorama==0.4.6
coloredlogs==15.0.1
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
humanfriendly==10.0
intel-openmp==2021.4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onnxruntime==1.18.1
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
protobuf==5.27.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sympy==1.12.1
tbb==2021.13.0
torch==2.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
